move "stdlib" and change borrows (unsure why)
